requirements.txt
amino.py
requests
